requirements.txt
absl-py==1.2.0
accelerate==0.15.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beartype==0.9.1
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.4
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colored==1.4.3
contourpy==1.0.7
copulas==0.8.0
crcmod==1.7
cryptography==3.4.8
ctgan==0.7.0
cycler==0.11.0
Cython==0.29.32
d4rl @ git+https://github.com/rail-berkeley/d4rl@d842aa194b416e564e54b0730d9f934e3e32f854
debugpy==1.5.1
decorator==5.1.1
deepecho==0.4.0
defusedxml==0.7.1
Deprecated==1.2.13
diffusers==0.10.2
dm-control @ git+https://github.com/deepmind/dm_control@f4e5e2336991017361f563db7e676fd08857076f
dm-env==1.5
dm-tree==0.1.7
dmcgym @ git+https://github.com/conglu1997/dmcgym.git@812905790dd87a448c9544a0beccb8b05ea2a850
docker-pycreds==0.4.0
einops==0.4.1
einops-exts==0.0.4
ema-pytorch==0.1.2
entrypoints==0.4
executing==0.8.3
Faker==14.2.1
fasteners==0.17.3
fastjsonschema==2.16.2
filelock==3.8.2
flax==0.4.0
flit_core==3.6.0
fonttools==4.38.0
future==0.18.2
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.4
google-auth==2.15.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.47.0
gym==0.18.0
gym-notices==0.0.8
h5py==3.7.0
huggingface-hub==0.11.1
humanfriendly==10.0
humanize==3.13.1
idna==3.4
imageio==2.9.0
importlib-metadata==5.1.0
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jax==0.2.21
jaxlib==0.3.15
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyter-server==1.18.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
labmaze==1.0.5
lockfile==0.12.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==0.8.4
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.4
mujoco==2.2.1
mujoco-py==2.1.2.14
mypy-extensions==0.4.3
nb-conda-kernels==2.3.1
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==22.1.2
pkgutil_resolve_name==1.3.10
plotly==5.13.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser==2.21
pyglet==1.5.0
Pygments==2.11.2
PyOpenGL==3.1.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrallis==0.3.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
ray==1.9.1
rdt==1.3.0
redis==4.3.4
-e git+https://github.com/watchernyu/REDQ.git@3a2763889223677d74c734864244c41494e4832c#egg=redq
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rotary-embedding-torch==0.2.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.2.0
scikit-video==1.1.11
scipy==1.9.0
sdmetrics==0.9.0
sdv==0.18.0
Send2Trash==1.8.0
sentry-sdk==1.9.5
setproctitle==1.3.2
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
slurm-gpustat==0.0.12
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
structlog==22.3.0
tenacity==8.2.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.13.1
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.2.1
toolz==0.12.0
torch==1.12.1+cu113
torchdiffeq==0.2.3
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typed-argument-parser==1.7.2
typing_extensions==4.4.0
typing-inspect==0.8.0
urllib3==1.26.13
wandb==0.13.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.1
zipp==3.8.1
